+++ mopack/origins/conan.py
@@ -15,7 +15,7 @@ from ..shell import ShellArguments
 
 class ConanPackage(BinaryPackage, BatchPackage):
     origin = 'conan'
-    _version = 1
+    _version = 2
 
     @GenericFreezeDried.fields(rehydrate={'extra_args': ShellArguments})
     class Options(PackageOptions):
@@ -41,6 +41,15 @@ class ConanPackage(BinaryPackage, BatchPackage):
 
     @staticmethod
     def upgrade(config, version):
+        # v2 moves `submodules.required` to `submodule_required` and stores
+        # `submodules` as a dict of submodule names.
+        if version < 2:
+            if config['submodules']:
+                config['submodule_required'] = config['submodules']['required']
+                config['submodules'] = {
+                    i: {} for i in config['submodules']['names']
+                }
+
         return config
 
     # TODO: Remove `usage` after v0.2 is released.