+++ mopack/origins/sdist.py
@@ -7,14 +7,14 @@ from typing import Dict, List
 from urllib.request import urlopen
 from subprocess import SubprocessError
 
-from . import Package, submodules_type
+from . import Package, submodules_type, submodule_required_type
 from .. import archive, log, types
 from ..builders import Builder, make_builder
 from ..config import ChildConfig
 from ..environment import get_cmd
 from ..freezedried import GenericFreezeDried
 from ..glob import filter_glob
-from ..iterutils import flatten, isiterable, listify
+from ..iterutils import flatten, isiterable, ismapping, listify
 from ..linkages import make_linkage
 from ..log import LogFile
 from ..objutils import Unset
@@ -32,10 +32,21 @@ from ..yaml_tools import to_parse_error
 }, skip_compare={'pkg_default', 'pending_builders', 'pending_linkage'})
 class SDistPackage(Package):
     # TODO: Remove `usage` after v0.2 is released.
-    def __init__(self, name, *, env=None, submodules=Unset, build=Unset,
-                 linkage=Unset, usage=Unset, inherit_defaults=False, _options,
-                 **kwargs):
-        if linkage is Unset and usage is not Unset:
+    def __init__(self, name, *, env=None, submodules=Unset,
+                 submodule_required=Unset, build=Unset, linkage=Unset,
+                 usage=Unset, inherit_defaults=False, _options, **kwargs):
+        # TODO: Remove this after v0.2 is released.
+        if ( ismapping(submodules) and
+             set(submodules.keys()) == {'names', 'required'} ):
+            warnings.warn(types.FieldKeyWarning(
+                ('`submodules` now takes a dictionary of submodules; use ' +
+                 '`submodule_required` to set whether submodules are ' +
+                 'required instead'), 'submodules'
+            ))
+            require_submodule = submodules.get('required', True)
+            submodules = {i: {} for i in submodules['names']}
+
+        if linkage is None and usage is not Unset:
             warnings.warn(types.FieldKeyWarning(
                 '`usage` is deprecated; use `linkage` instead', 'usage'
             ))
@@ -55,6 +66,9 @@ class SDistPackage(Package):
         self._expr_symbols = self._expr_symbols.augment(env=self.env)
         T = types.TypeCheck(locals(), self._expr_symbols)
         T.submodules(submodules_type(raw=True))
+        T.submodule_required(submodule_required_type(
+            self.submodules, raw=True
+        ))
 
         self.pending_builders = build
         self.pending_linkage = linkage
@@ -182,11 +196,23 @@ class SDistPackage(Package):
                     yield
 
         T = types.TypeCheck(export, dest=self)
+        recheck = (self.submodules is Unset and
+                   self.submodule_required is not Unset)
         with load_config(export.config_file, export.data):
             if self.submodules is Unset:
                 T.submodules(self.pkg_default(
                     submodules_type(), default=Unset
                 ))
+            if self.submodule_required is Unset:
+                T.submodule_required(self.pkg_default(
+                    submodule_required_type(self.submodules), default=Unset
+                ))
+
+        if recheck:
+            # Re-check the `submodule_required` field for consistency.
+            submodule_required_type(self.submodules)(
+                'submodule_required', self.submodule_required
+            )
 
         # TODO: Support package defaults for builders/linkage?
 
@@ -252,7 +278,7 @@ class SDistPackage(Package):
 @GenericFreezeDried.fields(rehydrate={'path': Path})
 class DirectoryPackage(SDistPackage):
     origin = 'directory'
-    _version = 3
+    _version = 4
 
     @staticmethod
     def upgrade(config, version):
@@ -263,6 +289,15 @@ class DirectoryPackage(SDistPackage):
         # v3 adds the `env` field.
         if version < 3:  # pragma: no branch
             config['env'] = {}
+
+        # v4 moves `submodules.required` to `submodule_required` and stores
+        # `submodules` as a dict of submodule names.
+        if version < 4:
+            if config['submodules']:
+                config['submodule_required'] = config['submodules']['required']
+                config['submodules'] = {
+                    i: {} for i in config['submodules']['names']
+                }
 
         return config
 
@@ -285,7 +320,7 @@ class DirectoryPackage(SDistPackage):
                            skip_compare={'guessed_srcdir'})
 class TarballPackage(SDistPackage):
     origin = 'tarball'
-    _version = 3
+    _version = 4
 
     @staticmethod
     def upgrade(config, version):
@@ -296,6 +331,15 @@ class TarballPackage(SDistPackage):
         # v3 adds the `env` field.
         if version < 3:  # pragma: no branch
             config['env'] = {}
+
+        # v4 moves `submodules.required` to `submodule_required` and stores
+        # `submodules` as a dict of submodule names.
+        if version < 4:
+            if config['submodules']:
+                config['submodule_required'] = config['submodules']['required']
+                config['submodules'] = {
+                    i: {} for i in config['submodules']['names']
+                }
 
         return config
 
@@ -384,7 +428,7 @@ class TarballPackage(SDistPackage):
 
 class GitPackage(SDistPackage):
     origin = 'git'
-    _version = 3
+    _version = 4
 
     @staticmethod
     def upgrade(config, version):
@@ -395,6 +439,15 @@ class GitPackage(SDistPackage):
         # v3 adds the `env` field.
         if version < 3:  # pragma: no branch
             config['env'] = {}
+
+        # v4 moves `submodules.required` to `submodule_required` and stores
+        # `submodules` as a dict of submodule names.
+        if version < 4:
+            if config['submodules']:
+                config['submodule_required'] = config['submodules']['required']
+                config['submodules'] = {
+                    i: {} for i in config['submodules']['names']
+                }
 
         return config