+++ test/unit/linkages/__init__.py
@@ -2,6 +2,7 @@ import os
 
 from .. import OptionsTest, MockPackage, through_json  # noqa: F401
 
+from mopack.iterutils import slice_dict
 from mopack.metadata import Metadata
 
 
@@ -15,7 +16,7 @@ class LinkageTest(OptionsTest):
         self.metadata = Metadata(self.pkgdir)
 
     def make_linkage(self, *args, common_options=None, deploy_dirs=None,
-                     submodules=None, **kwargs):
+                     **kwargs):
         if len(args) == 1:
             linkage_type = self.linkage_type
             pkg = args[0]
@@ -23,8 +24,11 @@ class LinkageTest(OptionsTest):
             linkage_type, pkg = args
 
         if isinstance(pkg, str):
+            pkg_kwargs = slice_dict(kwargs, {
+                'submodules', 'submodule_required'
+            })
             options = self.make_options(common_options, deploy_dirs)
             pkg = MockPackage(pkg, srcdir=self.srcdir, builddir=self.builddir,
-                              submodules=submodules, _options=options)
+                              _options=options, **pkg_kwargs)
 
         return linkage_type(pkg, _symbols=pkg._linkage_expr_symbols, **kwargs)