+++ test/unit/origins/sdist/test_git.py
@@ -351,11 +351,8 @@ class TestGit(SDistTestCase):
             mrun.assert_not_called()
 
     def test_submodules(self):
-        submodules_required = {'names': '*', 'required': True}
-        submodules_optional = {'names': '*', 'required': False}
-
         pkg = self.make_package('foo', repository=self.srcssh, build='bfg9000',
-                                submodules=submodules_required)
+                                submodules='*', submodule_required=True)
         self.check_fetch(pkg)
         self.check_resolve(pkg)
         self.check_linkage(pkg, submodules=['sub'])
@@ -363,7 +360,7 @@ class TestGit(SDistTestCase):
         pkg = self.make_package(
             'foo', repository=self.srcssh, build='bfg9000',
             linkage={'type': 'pkg_config', 'pcname': 'bar'},
-            submodules=submodules_required
+            submodules='*', submodule_required=True
         )
         self.check_fetch(pkg)
         self.check_resolve(pkg)
@@ -374,7 +371,7 @@ class TestGit(SDistTestCase):
         })
 
         pkg = self.make_package('foo', repository=self.srcssh, build='bfg9000',
-                                submodules=submodules_optional)
+                                submodules='*', submodule_required=False)
         self.check_fetch(pkg)
         self.check_resolve(pkg)
         self.check_linkage(pkg, submodules=['sub'])
@@ -382,7 +379,7 @@ class TestGit(SDistTestCase):
         pkg = self.make_package(
             'foo', repository=self.srcssh, build='bfg9000',
             linkage={'type': 'pkg_config', 'pcname': 'bar'},
-            submodules=submodules_optional
+            submodules='*', submodule_required=False
         )
         self.check_fetch(pkg)
         self.check_resolve(pkg)
@@ -393,9 +390,15 @@ class TestGit(SDistTestCase):
         })
 
     def test_invalid_submodule(self):
+        with self.assertRaises(TypeError):
+            self.make_package('foo', submodules={'sub': None},
+                              submodule_required=None)
+        with self.assertRaises(TypeError):
+            self.make_package('foo', submodule_required=True)
+
         pkg = self.make_package(
             'foo', repository=self.srcssh, build='bfg9000',
-            submodules={'names': ['sub'], 'required': True}, fetch=True
+            submodules={'sub': None}, submodule_required=True, fetch=True
         )
         with self.assertRaises(ValueError):
             pkg.get_linkage(self.metadata, ['invalid'])
@@ -665,9 +668,9 @@ class TestGit(SDistTestCase):
     def test_upgrade(self):
         opts = self.make_options()
         data = {
-            'origin': 'git', '_version': 0, 'name': 'foo',
+            'origin': 'git', '_version': 1, 'name': 'foo',
             'repository': 'repo', 'tag': None, 'branch': None, 'commit': None,
-            'srcdir': '.',
+            'srcdir': '.', 'submodules': {'names': ['sub'], 'required': False},
             'builder': {'type': 'none', '_version': 1, 'name': 'foo'},
             'linkage': {'type': 'system', '_version': 3},
         }
@@ -677,6 +680,8 @@ class TestGit(SDistTestCase):
             self.assertIsInstance(pkg, GitPackage)
             self.assertIsInstance(pkg.linkage, SystemLinkage)
             self.assertEqual([type(i) for i in pkg.builders], [NoneBuilder])
+            self.assertEqual(pkg.submodules, {'sub': {}})
+            self.assertEqual(pkg.submodule_required, False)
             m.assert_called_once()
 
     def test_builder_types(self):