+++ doc/about/changes.md
@@ -18,6 +18,8 @@ in progress
 - `custom` builders that use the `build/` output directory should set
   `outdir: build`
 - Linkages now use `submodule_linkage` instead of `submodule_map`
+- `submodules` field is now a dictionary of submodule names; to set the
+  `required` field, use `submodule_required` instead
 
 ---